requirements.txt
pywebview>=4.0.0